+++ scripts/gen-s-parser.py
@@ -43,15 +43,6 @@ instructions = [
     ("data.drop",      "makeDataDrop(s)"),
     ("memory.copy",    "makeMemoryCopy(s)"),
     ("memory.fill",    "makeMemoryFill(s)"),
-    ("i32.pop",        "makePop(Type::i32)"),
-    ("i64.pop",        "makePop(Type::i64)"),
-    ("f32.pop",        "makePop(Type::f32)"),
-    ("f64.pop",        "makePop(Type::f64)"),
-    ("v128.pop",       "makePop(Type::v128)"),
-    ("funcref.pop",    "makePop(Type::funcref)"),
-    ("externref.pop",  "makePop(Type::externref)"),
-    ("nullref.pop",    "makePop(Type::nullref)"),
-    ("exnref.pop",     "makePop(Type::exnref)"),
     ("i32.load",       "makeLoad(s, Type::i32, /*isAtomic=*/false)"),
     ("i64.load",       "makeLoad(s, Type::i64, /*isAtomic=*/false)"),
     ("f32.load",       "makeLoad(s, Type::f32, /*isAtomic=*/false)"),
@@ -503,7 +494,8 @@ instructions = [
     ("br_on_exn",            "makeBrOnExn(s)"),
     # Multivalue pseudoinstructions
     ("tuple.make",           "makeTupleMake(s)"),
-    ("tuple.extract",        "makeTupleExtract(s)")
+    ("tuple.extract",        "makeTupleExtract(s)"),
+    ("pop",                  "makePop(s)")
 ]