+++ src/wasm.h
@@ -62,6 +62,8 @@ struct Address {
   }
 };
 
+enum class IRProfile { Normal, Poppy };
+
 // Operators
 
 enum UnaryOp {
@@ -1095,6 +1097,8 @@ public:
   RefNull(MixedArena& allocator) {}
 
   void finalize();
+  void finalize(HeapType heapType);
+  void finalize(Type type);
 };
 
 class RefIsNull : public SpecificExpression<Expression::RefIsNullId> {
@@ -1259,7 +1263,8 @@ using StackIR = std::vector<StackInst*>;
 class Function : public Importable {
 public:
   Name name;
-  Signature sig;          // parameters and return value
+  Signature sig; // parameters and return value
+  IRProfile profile = IRProfile::Normal;
   std::vector<Type> vars; // non-param locals
 
   // The body of the function
@@ -1323,6 +1328,7 @@ public:
   Name getLocalNameOrGeneric(Index index);
 
   bool hasLocalName(Index index) const;
+  void setLocalName(Index index, Name name);
 
   void clearNames();
   void clearDebugInfo();
@@ -1496,6 +1502,9 @@ public:
   // too.
   FeatureSet features = FeatureSet::MVP;
   bool hasFeaturesSection = false;
+
+  // Module name, if specified. Serves a documentary role only.
+  Name name;
 
   MixedArena allocator;